Split main file into two
